Move QueueBuilder to a separate module
