+++ omniqueue/src/backends/rabbitmq.rs
@@ -15,12 +15,10 @@ pub use lapin::{
 };
 
 use crate::{
+    builder::{QueueBuilder, Static},
     decoding::DecoderRegistry,
     encoding::{CustomEncoder, EncoderRegistry},
-    queue::{
-        consumer::QueueConsumer, producer::QueueProducer, Acker, Delivery, QueueBackend,
-        QueueBuilder, Static,
-    },
+    queue::{consumer::QueueConsumer, producer::QueueProducer, Acker, Delivery, QueueBackend},
     scheduled::ScheduledProducer,
     QueueError,
 };
