+++ omniqueue/src/backends/redis/mod.rs
@@ -39,12 +39,10 @@ use svix_ksuid::KsuidLike;
 use tokio::task::JoinSet;
 
 use crate::{
+    builder::{QueueBuilder, Static},
     decoding::DecoderRegistry,
     encoding::{CustomEncoder, EncoderRegistry},
-    queue::{
-        consumer::QueueConsumer, producer::QueueProducer, Acker, Delivery, QueueBackend,
-        QueueBuilder, Static,
-    },
+    queue::{consumer::QueueConsumer, producer::QueueProducer, Acker, Delivery, QueueBackend},
     scheduled::ScheduledProducer,
     QueueError,
 };
