+++ omniqueue/tests/it/gcp_pubsub.rs
@@ -56,8 +56,11 @@ use google_cloud_pubsub::client::{Client, ClientConfig};
 use google_cloud_pubsub::subscription::SubscriptionConfig;
 use std::time::{Duration, Instant};
 
-use omniqueue::backends::gcp_pubsub::{GcpPubSubBackend, GcpPubSubConfig};
-use omniqueue::queue::{consumer::QueueConsumer, producer::QueueProducer, QueueBuilder, Static};
+use omniqueue::{
+    backends::gcp_pubsub::{GcpPubSubBackend, GcpPubSubConfig},
+    builder::{QueueBuilder, Static},
+    queue::{consumer::QueueConsumer, producer::QueueProducer},
+};
 use serde::{Deserialize, Serialize};
 
 const DEFAULT_PUBSUB_EMULATOR_HOST: &str = "localhost:8085";
